スタート画面とゲーム中画面にファイルを分けた
